BlogProduct/ProductApi/requirements.txt
django==5.0.6
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.15.1
pytest==8.2.1
pytest-django==4.8.0
django-mptt==0.16.0
drf-spectacular==0.27.2
coverage==7.5.3
pytest-cov==5.0.0
pytest-factoryboy==2.7.0